Add Loading SFX & Fix Bugs
The random repeating rate wasn't random, so I did the random repeating time inside a coroutine. The text would also repeat past 100% so I made a condition to check it doesn't repeat over the 100% limit. I then added some BG sound for the loading screen which plays when the loading screen is enabled.