Bugfixed Generator Meter & Cam Lag
In build, the progress bar in the PuzzleCanvas wouldn't decrease - this was because of issues with the approximate mathf function. I've replaced this to a simple bool which checks whether we are adding or subtracting to be specific. Camera lag with cinemachine was fixed by adding new input settings in the project settings.